add modals and alert service

+++ src/app/components/pages/users/users-page/users-list/users-list.component.ts
@@ -1,7 +1,9 @@
 import {ChangeDetectionStrategy, Component} from '@angular/core';
 import {Observable} from "rxjs";
-import {PaginatedUsersResponse, Users} from "../../../../../models/entities/User";
+import {PaginatedUsersResponse} from "../../../../../models/entities/User";
 import {UsersListService} from "./services/users-list.service";
+import {ModalService} from "../../../../../services/modals/modal.service";
+import {AlertService} from "../../../../../services/alerts/alert.service";
 
 @Component({
   selector: 'app-users-list',
@@ -12,7 +14,10 @@ import {UsersListService} from "./services/users-list.service";
 export class UsersListComponent {
   paginatedUsersResponse$: Observable<PaginatedUsersResponse>;
   currentUrl: string | null = null;
-  constructor(private _usersService: UsersListService) {
+  constructor(
+    private _usersService: UsersListService,
+    private _modalService: ModalService,
+  ) {
     this._usersService.getUsers();
     this.paginatedUsersResponse$ = this._usersService.paginatedUsersResponse$;
   }
@@ -23,7 +28,19 @@ export class UsersListComponent {
   }
 
   deleteUser(id: number): void {
-    this._usersService.deleteUser(id);
-    this._usersService.getUsers(this.currentUrl);
+    this._modalService.openModal({
+      title: 'Delete user',
+      body: 'Are you sure you want to delete this user?',
+      confirmButtonText: 'Delete',
+      confirmButtonClass: 'btn-success',
+      cancelButtonText: 'Cancel',
+      cancelButtonClass: 'btn-danger'
+    }).then((result: boolean):void => {
+      if (result) {
+        this._usersService.deleteUser(id);
+      }
+    }).catch((error: number): void => {
+      console.log('Error:', error);
+    });
   }
 }
